01-overview/Final Project/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
python-decouple==3.8
